fix: travis warnings

+++ src/gamelogic/WorldParser.hpp
@@ -22,9 +22,9 @@ namespace GameLogic {
         void fill_level(Level& level, std::string file_location);
     private:
         RapidXMLAdapter _rapid_adapter;
-        vector<vector<Tile*>> _generate_tilemap( vector<vector<size_t>> int_map, Engine::SPTR_AssetsManager assets );
-        CoordinateDouble _get_spawnpoint( vector<tuple<size_t, size_t, char*>> object_list, vector<vector<Tile*>> map );
-        void _set_objects( Level& level, vector<tuple<size_t, size_t, char*>> object_list, vector<vector<Tile*>> map, Engine::SPTR_AssetsManager assets );
+        vector<vector<Tile*>> _generate_tilemap( vector<vector<int>> int_map, Engine::SPTR_AssetsManager assets );
+        CoordinateDouble _get_spawnpoint( vector<tuple<int, int, char*>> object_list, vector<vector<Tile*>> map );
+        void _set_objects( Level& level, vector<tuple<int, int, char*>> object_list, vector<vector<Tile*>> map, Engine::SPTR_AssetsManager assets );
         double _spawn_tile_offset = 0.5;
         string _path;
     };
